refactored and improved conf module; 'xw::MainApplication' still need to be refactored

+++ src/controllers/controller.h
@@ -4,7 +4,6 @@
  * Copyright (c) 2019-2021 Yuriy Lisovskiy
  *
  * Intentionally simple parent class for all controllers.
- * Only implements dispatch-by-method and simple sanity checking.
  */
 
 #pragma once
@@ -14,7 +13,6 @@
 #include <functional>
 
 // Base libraries.
-//#include <xalwart.base/kwargs.h>
 #include <xalwart.base/result.h>
 #include <xalwart.base/utility.h>
 #include <xalwart.base/string_utils.h>
@@ -31,7 +29,7 @@
 __CONTROLLERS_BEGIN__
 
 template <typename ...ArgsT>
-using Handler = std::function<Result<std::shared_ptr<http::IHttpResponse>>(
+using Handler = std::function<http::result_t(
 	http::HttpRequest*, const std::tuple<ArgsT...>&, conf::Settings*
 )>;
 
@@ -40,12 +38,6 @@ using Handler = std::function<Result<std::shared_ptr<http::IHttpResponse>>(
 template <typename ...URLArgsT>
 class Controller
 {
-private:
-	inline Result<std::shared_ptr<http::IHttpResponse>> null()
-	{
-		return Result<std::shared_ptr<http::IHttpResponse>>::null();
-	}
-
 protected:
 	http::HttpRequest* request = nullptr;
 	conf::Settings* settings = nullptr;
@@ -55,56 +47,40 @@ protected:
 		"get", "post", "put", "patch", "delete", "head", "options", "trace"
 	};
 
-	// List of methods witch will be returned when OPTIONS is request.
+	// List of methods witch will be returned when 'OPTIONS' is in request.
 	std::vector<std::string> allowed_methods_list;
 
 protected:
 	inline explicit Controller(
 		const std::vector<std::string>& allowed_methods, conf::Settings* settings
-	) : Controller<URLArgsT...>(settings)
-	{
-		this->allowed_methods_list = allowed_methods;
-		if (std::find(allowed_methods.begin(), allowed_methods.end(), "options") == allowed_methods.end())
-		{
-			this->allowed_methods_list.emplace_back("options");
-		}
-	}
-
-	template <typename ResponseT, typename ...ArgsT>
-	inline Result<std::shared_ptr<http::IHttpResponse>> response(ArgsT&& ...args)
-	{
-		return Result<std::shared_ptr<http::IHttpResponse>>(
-				std::make_shared<ResponseT>(std::forward<ArgsT>(args)...)
-		);
-	}
-
-	template <unsigned short StatusCode, typename ...ArgsT>
-	inline Result<std::shared_ptr<http::IHttpResponse>> response(ArgsT&& ...args)
-	{
-		return Result<std::shared_ptr<http::IHttpResponse>>(
-			std::make_shared<http::HttpResponse>(StatusCode, std::forward<ArgsT>(args)...)
-		);
-	}
-
-	template <typename ...ArgsT>
-	inline Result<std::shared_ptr<http::IHttpResponse>> response(ArgsT&& ...args)
-	{
-		return Result<std::shared_ptr<http::IHttpResponse>>(std::forward<ArgsT>(args)...);
-	}
-
-	template<error_type ErrorType, typename ...ArgsT>
-	inline Result<std::shared_ptr<http::IHttpResponse>> raise(ArgsT&& ...args)
-	{
-		return raise<ErrorType, std::shared_ptr<http::IHttpResponse>>(std::forward<ArgsT>(args)...);
-	}
-
-public:
-	inline explicit Controller(conf::Settings* settings) : allowed_methods_list({"options"}), settings(settings)
+	) : settings(settings)
 	{
 		if (!this->settings)
 		{
 			throw NullPointerException("xw::ctrl::Controller: settings is nullptr", _ERROR_DETAILS_);
 		}
+
+		for (const auto& method : allowed_methods)
+		{
+			this->allowed_methods_list.push_back(str::lower(method));
+		}
+
+		auto options = std::find(this->allowed_methods_list.begin(), this->allowed_methods_list.end(), "options");
+		if (options == this->allowed_methods_list.end())
+		{
+			this->allowed_methods_list.emplace_back("options");
+		}
+	}
+
+	template<unsigned short int HttpErrorCode, typename ...ArgsT>
+	inline std::shared_ptr<http::IHttpResponse> raise(ArgsT&& ...args)
+	{
+		return std::make_shared<http::HttpResponse>(HttpErrorCode, std::forward<ArgsT>(args)...);
+	}
+
+public:
+	inline explicit Controller(conf::Settings* settings) : Controller<URLArgsT...>({"options"}, settings)
+	{
 	}
 
 	// Processes http GET request.
@@ -113,9 +89,9 @@ public:
 	// @param request: pointer to http request.
 	// @param args: pointer to requests' url arguments.
 	// @return pointer to http response instance.
-	virtual inline Result<std::shared_ptr<http::IHttpResponse>> get(URLArgsT ...args)
+	virtual inline http::result_t get(URLArgsT ...args)
 	{
-		return this->null();
+		return {};
 	}
 
 	// Processes http POST request.
@@ -124,9 +100,9 @@ public:
 	// @param request: pointer to http request.
 	// @param args: pointer to requests' url arguments.
 	// @return pointer to http response instance.
-	virtual inline Result<std::shared_ptr<http::IHttpResponse>> post(URLArgsT ...args)
+	virtual inline http::result_t post(URLArgsT ...args)
 	{
-		return this->null();
+		return {};
 	}
 
 	// Processes http PUT request.
@@ -135,40 +111,40 @@ public:
 	// @param request: pointer to http request.
 	// @param args: pointer to requests' url arguments.
 	// @return pointer to http response instance.
-	virtual inline Result<std::shared_ptr<http::IHttpResponse>> put(URLArgsT ...args)
+	virtual inline http::result_t put(URLArgsT ...args)
 	{
-		return this->null();
+		return {};
 	}
 
 	// Processes http PATCH request.
 	// Can be overridden in derived class, otherwise returns nullptr.
 	//
 	// @param request: pointer to http request.
-	// @param args: pointer to requests's url arguments.
+	// @param args: pointer to requests' url arguments.
 	// @return pointer to http response instance.
-	virtual inline Result<std::shared_ptr<http::IHttpResponse>> patch(URLArgsT ...args)
+	virtual inline http::result_t patch(URLArgsT ...args)
 	{
-		return this->null();
+		return {};
 	}
 
 	// Processes http DELETE request.
 	// Can be overridden in derived class, otherwise returns nullptr.
 	//
 	// @param request: pointer to http request.
-	// @param args: pointer to requests's url arguments.
+	// @param args: pointer to requests' url arguments.
 	// @return pointer to http response instance.
-	virtual inline Result<std::shared_ptr<http::IHttpResponse>> delete_(URLArgsT ...args)
+	virtual inline http::result_t delete_(URLArgsT ...args)
 	{
-		return this->null();
+		return {};
 	}
 
 	// Processes http HEAD request.
 	// Can be overridden in derived class, otherwise returns nullptr.
 	//
 	// @param request: pointer to http request.
-	// @param args: pointer to requests's url arguments.
+	// @param args: pointer to requests' url arguments.
 	// @return pointer to http response instance.
-	virtual inline Result<std::shared_ptr<http::IHttpResponse>> head(URLArgsT ...args)
+	virtual inline http::result_t head(URLArgsT ...args)
 	{
 		return this->get(args...);
 	}
@@ -177,29 +153,28 @@ public:
 	// Can be overridden in derived class, otherwise returns nullptr.
 	//
 	// @param request: pointer to http request.
-	// @param args: pointer to requests's url arguments.
+	// @param args: pointer to requests' url arguments.
 	// @return pointer to http response instance.
-	virtual inline Result<std::shared_ptr<http::IHttpResponse>> options(URLArgsT ...args)
+	virtual inline http::result_t options(URLArgsT ...args)
 	{
 		auto response = std::make_shared<http::HttpResponse>(200);
 		auto allowed_methods = this->allowed_methods();
 		response->set_header(
-			"Allow",
-			str::join(", ", allowed_methods.begin(), allowed_methods.end())
+			"Allow", str::join(", ", allowed_methods.begin(), allowed_methods.end())
 		);
 		response->set_header("Content-Length", "0");
-		return this->response(response);
+		return {response, nullptr};
 	}
 
 	// Processes http TRACE request.
 	// Can be overridden in derived class, otherwise returns nullptr.
 	//
 	// @param request: pointer to http request.
-	// @param args: pointer to requests's url arguments.
+	// @param args: pointer to requests' url arguments.
 	// @return pointer to http response instance.
-	virtual inline Result<std::shared_ptr<http::IHttpResponse>> trace(URLArgsT ...args)
+	virtual inline http::result_t trace(URLArgsT ...args)
 	{
-		return this->null();
+		return {};
 	}
 
 	// Setups request before dispatch call.
@@ -216,7 +191,7 @@ public:
 	//
 	// @param request: an actual http request from client.
 	// @return pointer to http response returned from handler.
-	virtual inline Result<std::shared_ptr<http::IHttpResponse>> dispatch(URLArgsT ...args)
+	virtual inline http::result_t dispatch(URLArgsT ...args)
 	{
 		if (this->request == nullptr)
 		{
@@ -229,43 +204,51 @@ public:
 		}
 
 		std::string method = str::lower(this->request->method());
-		auto result = Result<std::shared_ptr<http::IHttpResponse>>::null();
-		if (method == "get")
-		{
-			result = this->get(args...);
-		}
-		else if (method == "post")
-		{
-			result = this->post(args...);
-		}
-		else if (method == "put")
-		{
-			result = this->put(args...);
-		}
-		else if (method == "patch")
-		{
-			result = this->patch(args...);
-		}
-		else if (method == "delete")
-		{
-			result = this->delete_(args...);
-		}
-		else if (method == "head")
-		{
-			result = this->head(args...);
-		}
-		else if (method == "options")
-		{
-			result = this->options(args...);
-		}
-		else if (method == "trace")
-		{
-			result = this->trace(args...);
-		}
-
-		if (!result)
+		http::result_t result;
+		if (std::find(
+			this->allowed_methods_list.begin(), this->allowed_methods_list.end(), str::lower(method)
+		) == this->allowed_methods_list.end())
 		{
 			result = this->http_method_not_allowed();
+		}
+		else
+		{
+			if (method == "get")
+			{
+				result = this->get(args...);
+			}
+			else if (method == "post")
+			{
+				result = this->post(args...);
+			}
+			else if (method == "put")
+			{
+				result = this->put(args...);
+			}
+			else if (method == "patch")
+			{
+				result = this->patch(args...);
+			}
+			else if (method == "delete")
+			{
+				result = this->delete_(args...);
+			}
+			else if (method == "head")
+			{
+				result = this->head(args...);
+			}
+			else if (method == "options")
+			{
+				result = this->options(args...);
+			}
+			else if (method == "trace")
+			{
+				result = this->trace(args...);
+			}
+			else
+			{
+				result = this->http_method_not_allowed();
+			}
 		}
 
 		return result;
@@ -275,36 +258,31 @@ public:
 	//
 	// @param request: pointer to http request.
 	// @return pointer to http response returned from handler.
-	inline Result<std::shared_ptr<http::IHttpResponse>> http_method_not_allowed()
+	inline http::result_t http_method_not_allowed()
 	{
 		if (!this->settings->LOGGER)
 		{
-			throw NullPointerException("settings->logger is nullptr", _ERROR_DETAILS_);
+			throw NullPointerException("settings->LOGGER is nullptr", _ERROR_DETAILS_);
 		}
 
 		this->settings->LOGGER->warning(
 			"Method Not Allowed (" + this->request->method() + "): " + this->request->path(),
 			_ERROR_DETAILS_
 		);
-		return this->response<http::HttpResponseNotAllowed>("", this->allowed_methods());
+		return http::result<http::HttpResponseNotAllowed>("", this->allowed_methods());
 	}
 
-	// Builds vector of allowed methods.
-	// Used for http OPTIONS response.
-	// To make this method return correct allowed methods,
-	//		pass a vector of allowed methods names to protected
-	//		constructor in derived class initialization.
-	//
-	// @return std::vector of strings.
+	// Builds vector of allowed methods and used for http OPTIONS response.
+	// To make this method return correct allowed methods, pass a vector of
+	// allowed methods names to protected constructor in derived class
+	// initialization.
 	inline std::vector<std::string> allowed_methods()
 	{
 		std::vector<std::string> result;
 		for (const auto& method : this->allowed_methods_list)
 		{
 			bool found = std::find(
-				this->http_method_names.begin(),
-				this->http_method_names.end(),
-				str::lower(method)
+				this->http_method_names.begin(), this->http_method_names.end(), str::lower(method)
 			) != this->http_method_names.end();
 			if (found)
 			{
